requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-multipart==0.0.6
openai==1.60.2
pydantic==2.5.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
word2number==1.1
openpyxl==3.1.2
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
inflect==7.0.0
